pgorm-dev/requirements.txt
black==22.6.0
click==8.1.3
flake8==5.0.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
tomli==2.0.1
typing_extensions==4.3.0
